refactor: 포지션을 만들어내는 역할을 CellFinder에게 위임
